springboot-info-manage/src/test/java/com/fastcampus/javaallinone/project3/springbootinfomanage/service/PersonServiceTest.java
package com.fastcampus.javaallinone.project3.springbootinfomanage.service;

import com.fastcampus.javaallinone.project3.springbootinfomanage.domain.Block;
import com.fastcampus.javaallinone.project3.springbootinfomanage.domain.Person;
import com.fastcampus.javaallinone.project3.springbootinfomanage.repository.BlockRepository;
import com.fastcampus.javaallinone.project3.springbootinfomanage.repository.PersonRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.assertj.core.api.Assertions.assertThat;

import java.util.List;

@SpringBootTest
class PersonServiceTest {
    @Autowired
    private PersonService personService;
    @Autowired
    private PersonRepository personRepository;
    @Autowired
    private BlockRepository blockRepository;

    @Test
    void getPeopleExcludeBlocks(){
        List<Person> result = personService.getPeopleExcluedBlocks();

        result.forEach(System.out::println);

        assertThat(result.size()).isEqualTo(3);
        assertThat(result.get(0).getName()).isEqualTo("martin");
        assertThat(result.get(1).getName()).isEqualTo("david");
        assertThat(result.get(2).getName()).isEqualTo("benny");
    }

    @Test
    void getPoepleByName() {
        List<Person> result = personService.getPeopleByName("martin");

        assertThat(result.size()).isEqualTo(1);
        assertThat(result.get(0).getName()).isEqualTo("martin");
    }


    @Test
    void getPerson(){
        Person person = personService.getPerson(3L);

        assertThat(person.getName()).isEqualTo("dennis");
    }

}